# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HostYourHackathon API",
        "description": "REST API documentation for HostYourHackathon platform",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/judge/submissions": {
            "get": {
                "tags": [
                    "Judge"
                ],
                "summary": "Get assigned submissions for judging",
                "operationId": "judgeGetSubmissions",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/judge/submissions/{submission}/scores": {
            "put": {
                "tags": [
                    "Judge"
                ],
                "summary": "Update scores for a submission",
                "operationId": "judgeUpdateScores",
                "parameters": [
                    {
                        "name": "submission",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "criteria_id": {
                                            "type": "integer"
                                        },
                                        "score": {
                                            "type": "integer"
                                        },
                                        "remarks": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Judge"
                ],
                "summary": "Submit scores for a submission",
                "operationId": "judgeScoreSubmission",
                "parameters": [
                    {
                        "name": "submission",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "criteria_id": {
                                            "type": "integer"
                                        },
                                        "score": {
                                            "type": "integer"
                                        },
                                        "remarks": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/organizer/hackathons": {
            "post": {
                "tags": [
                    "Organizer"
                ],
                "summary": "Create a new hackathon",
                "operationId": "organizerCreateHackathon",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "allow_solo": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/organizer/hackathons/{hackathon}": {
            "put": {
                "tags": [
                    "Organizer"
                ],
                "summary": "Update a hackathon",
                "operationId": "organizerUpdateHackathon",
                "parameters": [
                    {
                        "name": "hackathon",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}